backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["gorilla97 <pranav.kosuri1@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.4"
requests = "^2.32.3"
uvicorn = "^0.32.0"
transformers = "^4.46.1"
pinecone-client = "^5.0.1"
neo4j = "^5.26.0"
python-dotenv = "^1.0.1"
pinecone = {extras = ["grpc"], version = "^5.3.1"}
pandas = "^2.2.3"
pydantic-settings = "^2.6.1"
yfinance = "^0.2.48"
pymupdf = "^1.24.13"
spacy = "^3.8.2"
openai = "^1.53.0"


[[tool.poetry.source]]
name = "pytorch_cpu"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
